# api/openapi.yml
openapi: 3.0.3
info:
  title: DormMate API (MVP)
  version: 0.1.0
  description: |
    DormMate MVP contract-first specification. 모든 코드는 이 명세와 동기화되어야 하며,
    CI에서는 런타임 OpenAPI와 본 파일을 비교해 드리프트를 감지한다.
servers:
  - url: http://localhost:8080
    description: 로컬 개발 기본 URL
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: 인증 실패
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: 권한 부족
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 처리 충돌
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    ReallocationPreviewRequest:
      type: object
      properties:
        floor:
          type: integer
          format: int32
          minimum: 1
      required: [floor]
    ReallocationPreviewResponse:
      type: object
      properties:
        floor:
          type: integer
          format: int32
        rooms:
          type: array
          items: { $ref: '#/components/schemas/ReallocationRoomSummary' }
        allocations:
          type: array
          items: { $ref: '#/components/schemas/ReallocationCompartmentAllocation' }
        chillCompartmentCount:
          type: integer
          format: int32
      required: [floor, rooms, allocations, chillCompartmentCount]
    ReallocationRoomSummary:
      type: object
      properties:
        roomId: { type: string, format: uuid }
        roomNumber: { type: string }
        roomType: { type: string }
        floor: { type: integer, format: int32 }
      required: [roomId, roomNumber, roomType, floor]
    ReallocationCompartmentAllocation:
      type: object
      properties:
        compartmentId: { type: string, format: uuid }
        fridgeUnitId: { type: string, format: uuid }
        slotIndex: { type: integer, format: int32 }
        slotLabel: { type: string }
        compartmentType:
          type: string
          enum: [CHILL, FREEZE]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, REPORTED, RETIRED]
        locked: { type: boolean }
        currentRoomIds:
          type: array
          items: { type: string, format: uuid }
        recommendedRoomIds:
          type: array
          items: { type: string, format: uuid }
        warnings:
          type: array
          items: { type: string }
      required:
        [compartmentId, fridgeUnitId, slotIndex, slotLabel, compartmentType, status, locked, currentRoomIds, recommendedRoomIds, warnings]
    ReallocationApplyRequest:
      type: object
      properties:
        floor:
          type: integer
          format: int32
          minimum: 1
        allocations:
          type: array
          items: { $ref: '#/components/schemas/ReallocationApplyAllocation' }
      required: [floor, allocations]
    ReallocationApplyAllocation:
      type: object
      properties:
        compartmentId: { type: string, format: uuid }
        roomIds:
          type: array
          items: { type: string, format: uuid }
      required: [compartmentId, roomIds]
    ReallocationApplyResponse:
      type: object
      properties:
        floor: { type: integer, format: int32 }
        affectedCompartments: { type: integer, format: int32 }
        releasedAssignments: { type: integer, format: int32 }
        createdAssignments: { type: integer, format: int32 }
        appliedAt: { type: string, format: date-time }
      required: [floor, affectedCompartments, releasedAssignments, createdAssignments, appliedAt]
    NotificationState:
      type: string
      enum: [UNREAD, READ, EXPIRED]
    NotificationItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        kindCode: { type: string }
        title: { type: string }
        body: { type: string }
        state: { $ref: '#/components/schemas/NotificationState' }
        createdAt: { type: string, format: date-time }
        readAt:
          type: string
          format: date-time
          nullable: true
        ttlAt:
          type: string
          format: date-time
          nullable: true
        correlationId:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
      required: [id, kindCode, title, body, state, createdAt]
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/NotificationItem' }
        page: { type: integer, format: int32 }
        size: { type: integer, format: int32 }
        totalElements: { type: integer, format: int64 }
        unreadCount: { type: integer, format: int64 }
      required: [items, page, size, totalElements, unreadCount]
    NotificationPreferenceItem:
      type: object
      properties:
        kindCode: { type: string }
        displayName: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        allowBackground: { type: boolean }
      required: [kindCode, displayName, description, enabled, allowBackground]
    NotificationPreferenceResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/NotificationPreferenceItem' }
      required: [items]
    UpdateNotificationPreferenceRequest:
      type: object
      properties:
        enabled: { type: boolean }
        allowBackground: { type: boolean }
      required: [enabled, allowBackground]
...
paths:
...
  /admin/fridge/reallocations/preview:
    post:
      tags: [FridgeAdmin]
      summary: 층별 칸-호실 재배분 추천안 조회
      description: 현재 배정 현황을 바탕으로 균등 분배한 추천안을 생성합니다. 실제 반영 전 검토용으로 사용하세요.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReallocationPreviewRequest' }
      responses:
        '200':
          description: 추천안 생성 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReallocationPreviewResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/fridge/reallocations/apply:
    post:
      tags: [FridgeAdmin]
      summary: 층별 칸-호실 재배분 확정 반영
      description: 검토·수정한 배정안을 저장하고 기존 배정을 종료합니다.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReallocationApplyRequest' }
      responses:
        '200':
          description: 재배분 적용 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReallocationApplyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications:
    get:
      tags: [Notifications]
      summary: 내 알림 목록 조회
      description: 미읽음 우선 정렬로 알림을 반환합니다. TTL이 지난 알림은 자동으로 만료 처리됩니다.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: state
          schema:
            type: string
            enum: [all, unread, read]
            default: all
          description: 조회할 알림 상태
        - in: query
          name: page
          schema: { type: integer, format: int32, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, format: int32, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: 알림 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: 특정 알림 읽음 처리
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: notificationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: 읽음 처리 완료
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /notifications/read-all:
    patch:
      tags: [Notifications]
      summary: 모든 미읽음 알림 읽음 처리
      security: [{ BearerAuth: [] }]
      responses:
        '204':
          description: 처리 완료
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /notifications/preferences:
    get:
      tags: [Notifications]
      summary: 알림 수신 설정 조회
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: 알림 선호 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferenceResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /notifications/preferences/{kindCode}:
    patch:
      tags: [Notifications]
      summary: 알림 수신 설정 변경
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: kindCode
          description: "알림 종류 코드 (예: FRIDGE_RESULT)"
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateNotificationPreferenceRequest' }
      responses:
        '200':
          description: 변경된 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferenceItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...